Add physic

+++ SpaghettiEngine/Framework/PhysicEngine/Shapes/Polygon.h
@@ -33,8 +33,8 @@ public:
 
 	void UpdateParameter() override;
 protected:
-	bool CheckCollideOnEachEdge(const Polygon& other, float& penatration, Vector3& normal);
-	bool CheckCollideOnOneEdge(const Vector3& edge, const Polygon& other, float& edgePenatration, Vector3& edgeNormal);
+	bool CheckCollideOnEachEdge(const Polygon& other, double& penatration, Vector3& normal);
+	bool CheckCollideOnOneEdge(const Vector3& edge, const Polygon& other, double& edgePenatration, Vector3& edgeNormal);
 protected:
 	std::vector<Vector3> _vertexes;
 	std::vector<Vector3> _worldVertexes;
